Make stories and components for prices
