+++ src/components/trade/Price.tsx
@@ -11,12 +11,14 @@ import { DEFAULT_PRECISION, MEDIA } from 'const'
 // Components
 import { OrderBookBtn } from 'components/OrderBookBtn'
 
-// TradeWidget: subcomponents
-import { TradeFormData } from 'components/TradeWidget'
+// Common
 import { FormInputError } from 'components/common/FormInputError'
 import { FormMessage } from 'components/common/FormMessage'
+
+// TradeWidget: subcomponents
+import { TradeFormData } from 'components/TradeWidget'
 import { useNumberInput } from 'components/TradeWidget/useNumberInput'
-import { SwapPrice } from './PriceEstimations'
+import { SwapPrice } from 'components/common/SwapPrice'
 
 const Wrapper = styled.div`
   display: flex;
@@ -60,7 +62,7 @@ const Wrapper = styled.div`
   }
 `
 
-export const PriceInputBox = styled.div<{ hidden?: boolean }>`
+const PriceInputBox = styled.div<{ hidden?: boolean }>`
   display: ${(props): string => (props.hidden ? 'none' : 'flex')};
   flex-flow: column nowrap;
   margin: 0;
@@ -156,7 +158,7 @@ export const PriceInputBox = styled.div<{ hidden?: boolean }>`
   }
 `
 
-interface Props {
+export interface Props {
   sellToken: TokenDetails
   receiveToken: TokenDetails
   priceInputId: string
@@ -176,7 +178,7 @@ export function invertPriceFromString(priceValue: string): string {
   return invertedPrice.isFinite() ? invertedPrice.toString(10) : '0'
 }
 
-const Price: React.FC<Props> = ({
+export const Price: React.FC<Props> = ({
   sellToken,
   receiveToken,
   priceInputId,
